[avr] Set RF69 OokFixedThresh on init which sets the noise floor in peak mode
I've experimented with dynamically adjusting the floor value, but there
doesn't seem to be much to gain from this. To determine if the floor is too
low, hook an oscilloscope to the data pin. There should be the occasional
pulse (every few seconds) indicating OOK receive. If it is too low it will
look like it is constantly receiving.

The floor could be set lower to increase sensitivity but a corresponding
decrease in receiver bandwidth should probably be used as well to increase
the selectivity.